+++ crates/wastebin_core/src/db.rs
@@ -254,6 +254,8 @@ pub mod read {
         pub title: Option<String>,
         /// Entry expiration datetime
         pub expiration: Option<Expiration>,
+        /// Entry will be deleted the next time it is fetched via [`Database::get`].
+        pub must_be_deleted: bool,
     }
 
     /// Potentially deleted or non-existent expired entry.
@@ -474,7 +476,7 @@ impl Handler {
 
     fn get_metadata(&self, id: Id) -> Result<Metadata, Error> {
         let metadata = self.conn.query_row(
-            "SELECT uid, title, CAST(ROUND((julianday(expires) - julianday('now')) * 86400) AS INTEGER) FROM entries WHERE id=?1",
+            "SELECT uid, title, CAST(ROUND((julianday(expires) - julianday('now')) * 86400) AS INTEGER), burn_after_reading FROM entries WHERE id=?1",
             params![id.to_i64()],
             |row| {
                 let expiration = row.get::<_, Option<i64>>(2)?
@@ -486,6 +488,7 @@ impl Handler {
                     uid: row.get(0)?,
                     title: row.get::<_, Option<String>>(1)?,
                     expiration,
+                    must_be_deleted: row.get::<_, Option<bool>>(3)?.unwrap_or(false),
                 })
             }
         )?;